♻️ コードレビュー指摘の修正・ETA機能の仕様除去
- TouringEtaWidget / TouringWeatherWidget を削除（TouringDestinationWidgetに統合済みで不要）
- /api/routing プロキシ（OSRM）を削除（ETA機能を仕様から除去）
- weatherEmoji 関数を lib/utils/weatherUtils.ts に共通化
- TouringModeView から未使用の endDate prop を削除
- E2Eヘルパー（startTestTouring/endTestTouring）を helpers/touringHelper.ts に移動
- TouringDestinationWidget にスポット名を表示する spotName prop を追加
