--- apps/web/components/touring/TouringEtaWidget.module.css
@@ -1,79 +1,0 @@
-.widget {
-  padding: var(--spacing-3) var(--spacing-4);
-  background-color: var(--card-surface);
-  border: 1px solid var(--card-border);
-  border-radius: var(--radius-md);
-}
-
-.header {
-  font-size: var(--font-size-xs);
-  color: var(--card-text-dim);
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  margin-bottom: var(--spacing-2);
-}
-
-.etaText {
-  font-size: var(--font-size-base);
-  font-weight: var(--font-weight-semibold);
-  color: var(--card-text);
-}
-
-.subText {
-  font-size: var(--font-size-xs);
-  color: var(--card-text-muted);
-  margin-top: 0.125rem;
-}
-
-.hint {
-  font-size: var(--font-size-sm);
-  color: var(--card-text-dim);
-}
-
-.loading {
-  font-size: var(--font-size-sm);
-  color: var(--card-text-dim);
-}
-
-.error {
-  font-size: var(--font-size-sm);
-  color: var(--card-text-dim);
-  opacity: 0.7;
-}
-
-.arrivedWidget {
-  border-color: var(--color-product);
-  background-color: color-mix(in srgb, var(--color-product) 8%, var(--card-surface));
-}
-
-.arrivedText {
-  font-size: var(--font-size-lg);
-  font-weight: var(--font-weight-bold);
-  color: var(--color-product);
-}
-
-.arrivalButton {
-  margin-top: var(--spacing-2);
-  width: 100%;
-  padding: var(--spacing-1) var(--spacing-2);
-  font-size: var(--font-size-xs);
-  font-weight: var(--font-weight-medium);
-  border-radius: var(--radius-md);
-  background-color: transparent;
-  border: 1px solid var(--card-border);
-  color: var(--card-text-muted);
-  cursor: pointer;
-  transition:
-    background-color 0.15s,
-    border-color 0.15s;
-}
-
-.arrivalButton:hover:not(:disabled) {
-  background-color: var(--color-cloudHover);
-  border-color: var(--color-cloudActive);
-}
-
-.arrivalButton:disabled {
-  opacity: 0.4;
-  cursor: not-allowed;
-}
