--- apps/web/components/touring/TouringWeatherWidget.module.css
@@ -1,60 +1,0 @@
-.widget {
-  padding: var(--spacing-3) var(--spacing-4);
-  background-color: var(--card-surface);
-  border: 1px solid var(--card-border);
-  border-radius: var(--radius-md);
-}
-
-.header {
-  font-size: var(--font-size-xs);
-  color: var(--card-text-dim);
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  margin-bottom: var(--spacing-2);
-}
-
-.content {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-3);
-}
-
-.weatherIcon {
-  font-size: 2rem;
-  line-height: 1;
-  flex-shrink: 0;
-}
-
-.details {
-  display: flex;
-  flex-direction: column;
-  gap: 0.125rem;
-}
-
-.temperature {
-  font-size: var(--font-size-xl);
-  font-weight: var(--font-weight-bold);
-  color: var(--card-text);
-  line-height: 1;
-}
-
-.description {
-  font-size: var(--font-size-sm);
-  color: var(--card-text-muted);
-}
-
-.subInfo {
-  font-size: var(--font-size-xs);
-  color: var(--card-text-dim);
-}
-
-.loading {
-  font-size: var(--font-size-sm);
-  color: var(--card-text-dim);
-}
-
-.error {
-  font-size: var(--font-size-sm);
-  color: var(--card-text-dim);
-  opacity: 0.7;
-}
