+++ apps/web/lib/utils/weatherUtils.ts
@@ -1,0 +1,12 @@
+/** WMO天気コードから絵文字への簡易マッピング */
+export function weatherEmoji(code: number): string {
+  if (code === 0) return '☀️'
+  if (code <= 2) return '🌤️'
+  if (code === 3) return '☁️'
+  if (code <= 48) return '🌫️'
+  if (code <= 67) return '🌧️'
+  if (code <= 77) return '❄️'
+  if (code <= 82) return '🌦️'
+  if (code <= 86) return '🌨️'
+  return '⛈️'
+}
